[c10d] Move pg wrapper tests to their own file.
moving these tests to their own standalone file. No meaningful code changes.

Differential Revision: [D29012664](https://our.internmc.facebook.com/intern/diff/D29012664/)

**NOTE FOR REVIEWERS**: This PR has internal Facebook specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D29012664/)!

[ghstack-poisoned]
