+++ test/distributed/test_c10d_common.py
@@ -233,125 +233,6 @@ class SparseGradientModule(nn.Module):
         return F.softmax(self.embedding(x), dim=1)
 
 
-class AbstractProcessGroupWrapperTest(MultiProcessTestCase):
-    def setUp(self):
-        super(AbstractProcessGroupWrapperTest, self).setUp()
-        # For Windows platform, Python does not support fork, change it to spawn here.
-        if sys.platform == "win32":
-            self._spawn_processes()
-        else:
-            self._fork_processes()
-
-    def _test_collective_hang(self, wrapper_pg, use_cuda=False):
-        # All ranks besides 1 call allreduce and wrapper_pg should detect a hang
-        # and report an issue with rank 1.
-        faulty_rank = 1
-        if self.rank != faulty_rank:
-            tensor = torch.randn(20, 10)
-            if use_cuda:
-                tensor = tensor.to(self.rank)
-
-            if self.rank == 0:
-                # Rank 0 reports faulty ranks
-                err = f"Ranks {faulty_rank} failed to pass monitoredBarrier"
-            else:
-                err = "Please check rank 0 logs for faulty rank"
-            with self.assertRaisesRegex(RuntimeError, err):
-                wrapper_pg.allreduce([tensor])
-
-    def _test_collectives_op_mismatch(self, wrapper_pg, use_cuda=False):
-        tensor = torch.randn(20, 10)
-        if use_cuda:
-            tensor = tensor.to(self.rank)
-        works = []
-        # Run a few successful collectives
-        for _ in range(10):
-            work = wrapper_pg.allreduce([tensor])
-            works.append(work)
-
-        for w in works:
-            w.wait()
-
-        # Simulate mismatch: allreduce vs reduce.
-        with self.assertRaisesRegex(
-            RuntimeError, "Mismatch between collective operation types"
-        ):
-            if self.rank == 0:
-                wrapper_pg.allreduce([tensor])
-            else:
-                wrapper_pg.reduce([tensor])
-
-        # Check additional mismatches
-
-        with self.assertRaisesRegex(
-            RuntimeError, "Mismatch between collective operation types"
-        ):
-            if self.rank == 0:
-                wrapper_pg.reduce([tensor])
-            else:
-                wrapper_pg.barrier()
-
-        with self.assertRaisesRegex(
-            RuntimeError, "Mismatch between collective operation types"
-        ):
-            scatter_result = [torch.ones(4) * i for i in range(self.world_size)]
-            scattered_tensor = torch.empty(4)
-            if self.rank == 0:
-                wrapper_pg.scatter(scattered_tensor, scatter_result, 0)
-            else:
-                wrapper_pg.reduce_scatter(scattered_tensor, scatter_result)
-
-        with self.assertRaisesRegex(
-            RuntimeError, "Mismatch between collective operation types"
-        ):
-            if self.rank == 0:
-                wrapper_pg.broadcast(tensor, 0)
-            else:
-                output_tensors = [
-                    torch.zeros_like(tensor) for _ in range(self.world_size)
-                ]
-                wrapper_pg.allgather([output_tensors], [tensor])
-
-    def _test_collective_shape_mismatch(self, wrapper_pg, use_cuda=False):
-        wrapper_pg.barrier()
-        dim = 2 if self.rank == 0 else 10
-        tensor = torch.randn(20, dim)
-        if use_cuda:
-            tensor = tensor.to(self.rank)
-        with self.assertRaisesRegex(RuntimeError, "Error when verifying shape tensors"):
-            wrapper_pg.allreduce([tensor])
-        # Check errors are raised when dimensionality of shapes is different
-        tensor = torch.randn(20, 10, 2) if self.rank == 0 else torch.randn(20, 10)
-        if use_cuda:
-            tensor = tensor.to(self.rank)
-        with self.assertRaisesRegex(RuntimeError, "Error when verifying shape tensors"):
-            wrapper_pg.allreduce([tensor])
-
-        # Check shape errors with scatter
-        input = [
-            torch.tensor(
-                [self.rank] if self.rank == 0 else [self.rank, self.rank],
-                device=self.rank if use_cuda else "cpu",
-            )
-            for _ in range(self.world_size)
-        ]
-        outputs = [
-            torch.tensor(
-                [-1] if self.rank == 0 else [-1, -1],
-                device=self.rank if use_cuda else "cpu",
-            )
-            for _ in range(self.world_size)
-        ]
-        root_rank = 0
-        opts = c10d.ScatterOptions()
-        opts.rootRank = root_rank
-        with self.assertRaisesRegex(RuntimeError, "Error when verifying shape tensors"):
-            if self.rank == root_rank:
-                wrapper_pg.scatter([outputs[self.rank]], [input], opts).wait()
-            else:
-                wrapper_pg.scatter([outputs[self.rank]], [], opts).wait()
-
-
 class AbstractDistributedDataParallelTest(object):
     def tearDown(self):
         # DistributedDataParallel test doesn't seem to call FileStore destructor
